Move engine with assets into separate namespace
Controller and assets are moved to separate namespace "restapi" so that it
doesn't colide with the host application. It also allows better sepraration of
assets for the host application and this engine.

With this change an option to determine the path to this engine was moved from
routing to configuration `baseurl` and a helper method `full_url` was
introduced and it's used to generate the url according to current settings.

New config attribute `baseurl` was introduced to make more accessble the
information about url the service is available on. Static content is
provided by a custom dispatcher that is able to take this parameter to
support another addresses for assets.

RAILS_RELATIVE_URL_ROOT is also supported with this patch.