Add ArrayOf validator
The intended usage

```ruby
param :ids, ArrayOfValidator.for_class, of: Integer
```

Explanation

Instead of simply checking the value is an Integer, it will try to find
a validator for that type.

Motivation

Rails of pass Integers as string. For example, instead of `ids: [1, 2]`
it could be `ids: ["1", "2"]`. In this case, we would still think this
is valid input.

But if we do

```ruby
param :ids, Array, of: Integer
```

an error will be raised for `"1"` is not an number