Fixes #629
The FunctionalTestRecording module's version of process() was not being called when running tests, which meant @collector was never initialized and so examples were not written to the file.
To fix this, we need to prepend FunctionalTestRecording to ActionController::TestCase::Behavior, not ActionController::TestCase, for reasons which I honestly did not manage to determine.