Allow custom validators to opt out of allow_blank behavior (#762)
The code that checks if fields are blank has a hard-coded carve-out for
the BooleanValidator so that false is an allowed value, even though it's
blank. Since this was hard-coded, there was not any way to allow blank
values for custom validators unless users define their parameters
specifically with allow_blank: true.

This commit adds a method to BaseValidator called "ignore_allow_blank?"
and updates the blank validation logic so that it no longer has a
hard-coded check for BooleanValidator and now just checks the
validator's value for "ignore_allow_blank?". The logic for
BooleanValidator is preserved, and custom validators can now also
achieve the same behavior.