Improve performance of route detection (#870)
In large apps with many routes (1500+), using the `api!` method
would be quite slow because `routes_for_action` looped over every route
trying to constantize the controller name. Compound this by the number of
times `api!` is used in the app.

Furthermore, each controller utilizing `api!` would constantize the next
controller using `api!`, which would constantize the next controller and so on,
recursively and potentially very deeply.

This change caches a hash of controller+action => routes using only the
controller name, avoiding the time-consuming loops, constantization, and recursion.

As a bonus, the cache is cleared when documentation is reloaded, fixing a bug where
new routes added while the app is running wouldn't get picked up.