Allow resource_name to be inherited (#872)
This is helpful if you have a very structured hierarchy of API controllers and would like to define the naming logic once in a base controller.

* Allow resource_name to be a string, symbol, or proc
* Search all the way up the controller hierarchy for a value
* Continue searching even if a particular controller doesn't have a `resource_description` block

This is helpful if you have a very structured hierarchy of API controllers and would like to define the naming logic once in a base controller.

* Fix whitespace on README

This was causing the test commands not to show up on Github

* Set Rubocop target Ruby version to 2.6

I had to set Rubocop's target Ruby version to 2.6 in order to get a bunch of offenses to go away. I'm not sure why that suddenly started happening (my PR from last week didn't have any issues).