Fix CI: Get the build green for modern ruby and rack and rubocop-rspec (#939)
* Update how we use Rack::Utils to get status codes

As part of https://github.com/rack/rack/pull/2137 (released in Rack 3) Rack
changed the values of `Rack::Utils::SYMBOL_TO_STATUS_CODE` so asking for
`:unprocessable_entity` now returns `nil`. The guidance has always been to use
`Rack::Util.status_code` (present in the API since Rack 1.1) so let's just do
that.

* Handle a missing `rack.input` in the env

Ever since https://github.com/rack/rack/pull/2018 `rack.input` has been
optional so we can't assume it's there and call `read` or `rewind` on it in
`Apipie::Extractor::Recorder`. Using safe navigation to call these methods
seems like the simplest approach because we want to look for
`rack.request.form_hash` instead in both the situation where `rack.input` is
missing, or it's empty.

* Add rubocop-rspec_rails and get rubocop run to green

In https://github.com/rubocop/rubocop-rspec/pull/1848 rubocop-rspec extracted
some cops to separate gems, and because we referenced one of them in our
rubocop_todo.yml the whole rubocop process wouldn't run because it doesn't like
config for cops it doesn't know about. We introducing the missing gem,
`rubocop-rspec_rails`, and fix the name in the config to let us run rubocop
again.

There is one infraction that exists in the source so we fix that to get us to a
green rubocop run.

* Only add rubocop-rspec_rails for ruby >-= 2.7

It's only available for 2.7+ and we still run this build on 2.6.