Handle warnings for param descriptions without a controller method
If we're generating swagger via
`SwaggerGenerator.json_schema_for_self_describing_class` we explicitly don't
have a `controller_method` being passed around so we can't use it for the
warnings. Introduce a test for type and builder to cover this scenario (first
spotted by a failing spec for `SwaggerGenerator`) and then change the
implementation to cope with it. Luckily,
`Apipie::Generator::Swagger::MethodDescription::Decorator` already had a
`ruby_name` implementation that handles being given `nil`, so we use that.